include directive inside module folder.
This example demonstrates that you could potentially
add files that export functions in the ng-module folder.
It is not recommended because you have to pick one scheme
for loading modules: either rely on angular module system
or rely on webpack to load the files for you. Again this
example exists to show that you could potentially load
some legacy modules if you need to.